+++ web/hooks/useSpacingStyles.ts
@@ -1,0 +1,76 @@
+import { useMemo } from 'react';
+
+import { SpacingScale } from '@cds/core';
+import { cx } from 'linaria';
+
+import * as marginStyles from '../styles/margin';
+import * as paddingStyles from '../styles/padding';
+import { isRtl } from '../utils/isRtl';
+
+type Props = Partial<
+  Readonly<{
+    all: SpacingScale;
+    bottom: SpacingScale;
+    end: SpacingScale;
+    horizontal: SpacingScale;
+    start: SpacingScale;
+    top: SpacingScale;
+    vertical: SpacingScale;
+    isInverted: boolean;
+  }>
+>;
+
+export const useSpacingStyles = ({
+  all,
+  bottom,
+  end,
+  horizontal,
+  start,
+  top,
+  vertical,
+  isInverted = false,
+}: Props): string =>
+  useMemo(() => {
+    const spacingClasses = [];
+    const styles = isInverted ? marginStyles : paddingStyles;
+
+    if (all) {
+      spacingClasses.push(styles.all[all]);
+    }
+
+    if (top) {
+      spacingClasses.push(styles.top[top]);
+    }
+
+    if (bottom) {
+      spacingClasses.push(styles.bottom[bottom]);
+    }
+
+    if (start) {
+      if (isRtl()) {
+        spacingClasses.push(styles.right[start]);
+      } else {
+        spacingClasses.push(styles.left[start]);
+      }
+    }
+
+    if (end) {
+      if (isRtl()) {
+        spacingClasses.push(styles.left[end]);
+      } else {
+        spacingClasses.push(styles.right[end]);
+      }
+    }
+
+    if (vertical) {
+      spacingClasses.push(styles.top[vertical]);
+      spacingClasses.push(styles.bottom[vertical]);
+    }
+
+    if (horizontal) {
+      spacingClasses.push(styles.left[horizontal]);
+      spacingClasses.push(styles.right[horizontal]);
+    }
+
+    return cx(...spacingClasses);
+  }, [all, top, bottom, start, end, vertical, horizontal, isInverted]);