+++ web/styles/margin.ts
@@ -1,0 +1,185 @@
+/**
+ * DO NOT MODIFY
+ * Generated from scripts/codegen/main.ts
+ */
+import { css } from 'linaria';
+
+export const all = {
+  1: css`
+    margin: -var(--spacing-1);
+  `,
+  2: css`
+    margin: -var(--spacing-2);
+  `,
+  3: css`
+    margin: -var(--spacing-3);
+  `,
+  4: css`
+    margin: -var(--spacing-4);
+  `,
+  5: css`
+    margin: -var(--spacing-5);
+  `,
+  6: css`
+    margin: -var(--spacing-6);
+  `,
+  7: css`
+    margin: -var(--spacing-7);
+  `,
+  8: css`
+    margin: -var(--spacing-8);
+  `,
+  9: css`
+    margin: -var(--spacing-9);
+  `,
+  10: css`
+    margin: -var(--spacing-10);
+  `,
+  0.5: css`
+    margin: -var(--spacing-0\\.5);
+  `,
+};
+
+export const top = {
+  1: css`
+    margin-top: -var(--spacing-1);
+  `,
+  2: css`
+    margin-top: -var(--spacing-2);
+  `,
+  3: css`
+    margin-top: -var(--spacing-3);
+  `,
+  4: css`
+    margin-top: -var(--spacing-4);
+  `,
+  5: css`
+    margin-top: -var(--spacing-5);
+  `,
+  6: css`
+    margin-top: -var(--spacing-6);
+  `,
+  7: css`
+    margin-top: -var(--spacing-7);
+  `,
+  8: css`
+    margin-top: -var(--spacing-8);
+  `,
+  9: css`
+    margin-top: -var(--spacing-9);
+  `,
+  10: css`
+    margin-top: -var(--spacing-10);
+  `,
+  0.5: css`
+    margin-top: -var(--spacing-0\\.5);
+  `,
+};
+
+export const bottom = {
+  1: css`
+    margin-bottom: -var(--spacing-1);
+  `,
+  2: css`
+    margin-bottom: -var(--spacing-2);
+  `,
+  3: css`
+    margin-bottom: -var(--spacing-3);
+  `,
+  4: css`
+    margin-bottom: -var(--spacing-4);
+  `,
+  5: css`
+    margin-bottom: -var(--spacing-5);
+  `,
+  6: css`
+    margin-bottom: -var(--spacing-6);
+  `,
+  7: css`
+    margin-bottom: -var(--spacing-7);
+  `,
+  8: css`
+    margin-bottom: -var(--spacing-8);
+  `,
+  9: css`
+    margin-bottom: -var(--spacing-9);
+  `,
+  10: css`
+    margin-bottom: -var(--spacing-10);
+  `,
+  0.5: css`
+    margin-bottom: -var(--spacing-0\\.5);
+  `,
+};
+
+export const left = {
+  1: css`
+    margin-left: -var(--spacing-1);
+  `,
+  2: css`
+    margin-left: -var(--spacing-2);
+  `,
+  3: css`
+    margin-left: -var(--spacing-3);
+  `,
+  4: css`
+    margin-left: -var(--spacing-4);
+  `,
+  5: css`
+    margin-left: -var(--spacing-5);
+  `,
+  6: css`
+    margin-left: -var(--spacing-6);
+  `,
+  7: css`
+    margin-left: -var(--spacing-7);
+  `,
+  8: css`
+    margin-left: -var(--spacing-8);
+  `,
+  9: css`
+    margin-left: -var(--spacing-9);
+  `,
+  10: css`
+    margin-left: -var(--spacing-10);
+  `,
+  0.5: css`
+    margin-left: -var(--spacing-0\\.5);
+  `,
+};
+
+export const right = {
+  1: css`
+    margin-right: -var(--spacing-1);
+  `,
+  2: css`
+    margin-right: -var(--spacing-2);
+  `,
+  3: css`
+    margin-right: -var(--spacing-3);
+  `,
+  4: css`
+    margin-right: -var(--spacing-4);
+  `,
+  5: css`
+    margin-right: -var(--spacing-5);
+  `,
+  6: css`
+    margin-right: -var(--spacing-6);
+  `,
+  7: css`
+    margin-right: -var(--spacing-7);
+  `,
+  8: css`
+    margin-right: -var(--spacing-8);
+  `,
+  9: css`
+    margin-right: -var(--spacing-9);
+  `,
+  10: css`
+    margin-right: -var(--spacing-10);
+  `,
+  0.5: css`
+    margin-right: -var(--spacing-0\\.5);
+  `,
+};