+++ web/utils/isRtl.ts
@@ -1,0 +1,3 @@
+export const isRtl = (element: HTMLElement = document.documentElement) => {
+  return element.dir === 'rtl';
+};