+++ core/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- change: Update spacing types by @hannah-jin [#11916](https://github.cbhq.net/mono/repo/pull/11916)
+
 ## 0.0.4 (February 1st, 2021 8:15am PST) [#12013](https://github.cbhq.net/mono/repo/pull/12013)
 
 - feat: Add Text shared props @hannah-jin ([#11435](https://github.cbhq.net/mono/repo/pull/11435)