+++ theme/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- feat: Add spacing variables by @hannah-jin [#11916](https://github.cbhq.net/mono/repo/pull/11916)
+
 ## 0.0.4 (January 25th, 2021 3:30pm PST) [#11711](https://github.cbhq.net/mono/repo/pull/11711)
 
 - chore: Update dependencies in package.json by @katherinemartinez [#11711](https://github.cbhq.net/mono/repo/pull/11711)